# server/src/handlers/data/openapi.yaml
openapi: 3.0.1
info:
  title: ORA View API
  version: 1.0.0
paths:
...
  /api/view/{uuid}:
    get:
      summary: Get node view
      operationId: getView
      tags:
        - View
      parameters:
        - name: uuid
          in: path
          required: true
          description: Node identifier.
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9_-]{1,128}$"
          example: "550e8400-e29b-41d4-a716-446655440000"
        - name: paths
          in: query
          required: false
          description: >
            RAMEN access paths whose values should be included in the response.
            A path can be a direct field, a nested access path, or a node access
            path ending with `*`.
          schema:
            type: array
            maxItems: 50
            items:
              $ref: "#/components/schemas/AccessPathString"
          style: form
          explode: true
          examples:
            directFields:
              value:
                - title
                - date
            nestedValues:
              value:
                - annotation[type].label
                - refers[Person].name
                - parents[Collection].title
                - children[Content].*
      responses:
        "200":
          description: Node view.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ViewNodeOnlyResponse"
                  - $ref: "#/components/schemas/ViewNodeWithValuesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
...
  responses:
    BadRequest:
      description: Invalid request parameters.
    NotFound:
      description: Resource not found.
    InternalError:
      description: Internal server error.
  schemas:
    AccessPathString:
      type: string
      minLength: 1
      maxLength: 512
      description: >
        RAMEN access path. Supported forms are direct field access such as
        `label`, nested access such as `annotation[type].label`,
        `refers[Person].name`, `parents[Collection].title`,
        `children[Content].label`, and node access such as `children[Content].*`.
        Supported access steps are `annotation`, `refers`, `parents`, and
        `children`.
      pattern: "^(?:(?:annotation|refers|parents|children)(?:\\[[^\\]]+\\])?\\.)*(?:[A-Za-z_]\\w*|\\*)$"
      examples:
        - label
        - annotation[type].label
        - refers[Person].name
        - parents[Collection].title
        - children[Content].*
...
    ViewNodeOnlyResponse:
      type: object
      required:
        - node
      properties:
        node:
          $ref: "#/components/schemas/Node"
    ViewNodeWithValuesResponse:
      type: object
      required:
        - node
        - values
      properties:
        node:
          $ref: "#/components/schemas/Node"
        values:
          type: object
          description: >
            Map of requested access path strings to their resolved values.
            A value can be a scalar, an object node, an array, or null-like
            JSON value depending on the requested path and database result.
          additionalProperties: true
          example:
            annotation[type].label: Place
            children[Content].*:
              _labels:
                - Content
              uuid: content-123
              label: Example content
    Node:
      type: object
      required:
        - uuid
        - _labels
      properties:
        uuid:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        _labels:
          type: array
          items:
            type: string
            enum:
              - Collection
              - Entity
              - Content
              - Annotation
          example:
            - Entity
      additionalProperties: true
      description: Neo4j node projection with all node properties and an additional `_labels` array.